backend/pyproject.toml
[tool.poetry]
name = "python-flask-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
sqlalchemy = "^2.0.27"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
notebook = "^7.1.0"
jupyter = "^1.0.0"
ipython = "^8.21.0"
ipykernel = "^6.29.2"
env = "^0.1.0"
pytest = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
